requirements.txt
flask
psycopg2-binary
cassandra-driver
python-dotenv